requirements.txt
Django==3.2
django-filter==21.1
django-widget-tweaks==1.4.12
django-taggit==2.1.0
python-decouple==3.6
dj-database-url==0.5.0
psycopg2==2.9.3
Pillow==9.1.0
django-crispy-forms==1.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-ckeditor==6.3.0
django-hitcount==1.3.5
fontawesomefree==6.1.1
GDAL==3.4.2
geoserver-rest==2.2.2
Shapely==1.8.1.post1
Fiona==1.8.21
pyproj==3.3.0
rtree==1.0.0
geopandas==0.10.2
rasterio==1.2.10
SQLAlchemy==1.4.35
